+++ src/Capability.php
@@ -11,7 +11,7 @@ namespace theantichris\SPF;
  * @package theantichris\SPF
  * @since 3.0.0
  */
-abstract class Capability extends Enum
+abstract class Capability
 {
     const switch_themes          = 'switch_themes';
     const edit_themes            = 'edit_themes';
@@ -71,4 +71,18 @@ abstract class Capability extends Enum
     const manage_network_users   = 'manage_network_users';
     const manage_network_themes  = 'manage_network_themes';
     const manage_network_options = 'manage_network_options';
+
+    /**
+     * Checks if a given value is a valid member of the fake enum.
+     *
+     * @since 3.0.0
+     *
+     * @param string $value
+     *
+     * @return bool
+     */
+    public static function isValid($value)
+    {
+        return defined("self::{$value}");
+    }
 }